move raw_data schema files to RawData folder
