+++ Data.SchemaMigrator/Models/PgContext/RawData/BassinVersantTopographique.cs
@@ -2,16 +2,14 @@
 using System.Collections.Generic;
 using NetTopologySuite.Geometries;
 
-namespace Data.SchemaMigrator.Models.PgContext
+namespace Data.SchemaMigrator.Models.PgContext.RawData
 {
-    public partial class SurfaceHydrographique
+    public partial class BassinVersantTopographique
     {
         public string Id { get; set; }
         public string CodeHydro { get; set; }
-        public string CodePays { get; set; }
-        public string Nature { get; set; }
-        public string PosSol { get; set; }
-        public string Etat { get; set; }
+        public string Toponyme { get; set; }
+        public string BassHydro { get; set; }
         public string DateCreat { get; set; }
         public string DateMaj { get; set; }
         public string DateApp { get; set; }
@@ -19,20 +17,14 @@ namespace Data.SchemaMigrator.Models.PgContext
         public string Source { get; set; }
         public string IdSource { get; set; }
         public string PrecPlani { get; set; }
-        public string PrecAlti { get; set; }
         public string SrcCoord { get; set; }
-        public string SrcAlti { get; set; }
         public string Statut { get; set; }
-        public string Persistanc { get; set; }
-        public string Salinite { get; set; }
+        public string BFluvial { get; set; }
         public string Origine { get; set; }
         public string Comment { get; set; }
-        public string IdPEau { get; set; }
+        public string CodeBh { get; set; }
+        public string CodeCarth { get; set; }
         public string IdCEau { get; set; }
-        public string IdEntTr { get; set; }
-        public string NomPEau { get; set; }
-        public string NomCEau { get; set; }
-        public string NomEntTr { get; set; }
         public Geometry Geometry { get; set; }
     }
 }
