+++ Data.SchemaMigrator/Models/PgContext/RawData/DetailHydrographique.cs
@@ -2,14 +2,17 @@
 using System.Collections.Generic;
 using NetTopologySuite.Geometries;
 
-namespace Data.SchemaMigrator.Models.PgContext
+namespace Data.SchemaMigrator.Models.PgContext.RawData
 {
-    public partial class BassinVersantTopographique
+    public partial class DetailHydrographique
     {
         public string Id { get; set; }
-        public string CodeHydro { get; set; }
+        public string Nature { get; set; }
+        public string NatDetail { get; set; }
         public string Toponyme { get; set; }
-        public string BassHydro { get; set; }
+        public string StatutTop { get; set; }
+        public string Importance { get; set; }
+        public string Etat { get; set; }
         public string DateCreat { get; set; }
         public string DateMaj { get; set; }
         public string DateApp { get; set; }
@@ -17,14 +20,6 @@ namespace Data.SchemaMigrator.Models.PgContext
         public string Source { get; set; }
         public string IdSource { get; set; }
         public string PrecPlani { get; set; }
-        public string SrcCoord { get; set; }
-        public string Statut { get; set; }
-        public string BFluvial { get; set; }
-        public string Origine { get; set; }
-        public string Comment { get; set; }
-        public string CodeBh { get; set; }
-        public string CodeCarth { get; set; }
-        public string IdCEau { get; set; }
         public Geometry Geometry { get; set; }
     }
 }
