+++ Data.SchemaMigrator/Models/PgContext/RawData/LimiteTerreMer.cs
@@ -2,17 +2,15 @@
 using System.Collections.Generic;
 using NetTopologySuite.Geometries;
 
-namespace Data.SchemaMigrator.Models.PgContext
+namespace Data.SchemaMigrator.Models.PgContext.RawData
 {
-    public partial class DetailHydrographique
+    public partial class LimiteTerreMer
     {
         public string Id { get; set; }
-        public string Nature { get; set; }
-        public string NatDetail { get; set; }
-        public string Toponyme { get; set; }
-        public string StatutTop { get; set; }
-        public string Importance { get; set; }
-        public string Etat { get; set; }
+        public string CodeHydro { get; set; }
+        public string CodePays { get; set; }
+        public string TypeLimit { get; set; }
+        public string Niveau { get; set; }
         public string DateCreat { get; set; }
         public string DateMaj { get; set; }
         public string DateApp { get; set; }
@@ -20,6 +18,10 @@ namespace Data.SchemaMigrator.Models.PgContext
         public string Source { get; set; }
         public string IdSource { get; set; }
         public string PrecPlani { get; set; }
+        public string SrcCoord { get; set; }
+        public string Statut { get; set; }
+        public string Origine { get; set; }
+        public string Comment { get; set; }
         public Geometry Geometry { get; set; }
     }
 }
