+++ Data.SchemaMigrator/Models/PgContext/RawData/Region.cs
@@ -2,13 +2,12 @@
 using System.Collections.Generic;
 using NetTopologySuite.Geometries;
 
-namespace Data.SchemaMigrator.Models.PgContext
+namespace Data.SchemaMigrator.Models.PgContext.RawData
 {
-    public partial class ArrondissementDepartemental
+    public partial class Region
     {
         public string Id { get; set; }
-        public string InseeArr { get; set; }
-        public string InseeDep { get; set; }
+        public string NomReg { get; set; }
         public string InseeReg { get; set; }
         public Geometry Geometry { get; set; }
     }
